# in -dtimings output, show time spent in C linker clearly