Before:

```
$ ./ocamlopt.opt -I otherlibs/unix -I stdlib -dtimings unix.cmxa -o foo
0.134s foo
  0.001s selection
  0.001s cse
  0.002s liveness
  0.001s spill
  0.001s split
  0.007s regalloc
  0.001s emit
  0.011s assemble
  0.109s other
0.008s other
```

After:
```
$ ./ocamlopt.opt -I otherlibs/unix -I stdlib -dtimings unix.cmxa -o foo
0.130s foo
  0.001s selection
  0.001s cse
  0.002s liveness
  0.001s spill
  0.001s split
  0.007s regalloc
  0.001s emit
  0.011s assemble
  0.100s c-linker
  0.004s other
0.006s other
```